Corrected type of a printk argument in resize_async_buffer().

+++ comedi/comedi_fops.c
@@ -2340,7 +2340,7 @@ static int resize_async_buffer(comedi_device *dev,
 	}
 
 	DPRINTK("comedi%i subd %d buffer resized to %i bytes\n",
-		dev->minor, s - dev->subdevices, async->prealloc_bufsz);
+		dev->minor, (int)(s - dev->subdevices), async->prealloc_bufsz);
 	return 0;
 }
